Subtitle work amb work
Working on getting subtitles scaling properly as well as being fully customizable for accessibility. This includes different colors for different speakers and the ability to add outline or adjust background opacity.

Also working on getting footsteps in fluid working. So far still not getting the line trace to connect with the water surface.